refactor: store in sessionstorage to remove items when browser or tab is closed; clear storage on signout
