+++ frontend/src/const.ts
@@ -1,6 +1,6 @@
 export const loginSessionStorageKey = 'userSession';
 
-export const CHAT_WIDGET_LOCALSTORAGE_KEYS = {
+export const CHAT_WIDGET_STORAGE_KEYS = {
 	privateTabs: 'privateTabs',
 	privateMessages: 'privateMessages',
 	generalMessages: 'generalMessages',
